+++ generators/java/sdk/src/main/java/com/fern/java/client/generators/visitors/FilePropertyIsOptional.java
@@ -1,8 +1,8 @@
 package com.fern.java.client.generators.visitors;
 
-import com.fern.irV42.model.http.FileProperty;
-import com.fern.irV42.model.http.FilePropertyArray;
-import com.fern.irV42.model.http.FilePropertySingle;
+import com.fern.ir.model.http.FileProperty;
+import com.fern.ir.model.http.FilePropertyArray;
+import com.fern.ir.model.http.FilePropertySingle;
 
 public class FilePropertyIsOptional implements FileProperty.Visitor<Boolean> {