+++ seed/ruby-model/undiscriminated-unions/lib/seed_undiscriminated_unions_client.rb
@@ -1,3 +1,6 @@
 # frozen_string_literal: true
 
 require_relative "seed_undiscriminated_unions_client/union/types/my_union"
+require_relative "seed_undiscriminated_unions_client/union/types/key_type"
+require_relative "seed_undiscriminated_unions_client/union/types/key"
+require_relative "seed_undiscriminated_unions_client/union/types/metadata"